+++ test/test-data/insert-element.ts
@@ -3,6 +3,7 @@ import { insertElement } from "../../src";
 
 const testObj1 = { id: "obj1", name: "foo" };
 const testObj2 = { id: "obj2", names: ["foo", "bar"] };
+const ROOT_QUERY = { products: [] };
 
 export const insertElementTestData: ReadonlyArray<OneTest> = [
   {
@@ -28,5 +29,30 @@ export const insertElementTestData: ReadonlyArray<OneTest> = [
     patches: [insertElement<typeof testObj2>("obj2", "names", 0, "baz")],
     cacheBefore: { obj2: { id: "obj2", names: null } },
     cacheAfter: { obj2: { id: "obj2", names: ["baz"] } }
+  },
+  {
+    name:
+      "insertElement with arguments should only insert in field with arguments",
+    patches: [
+      insertElement<typeof ROOT_QUERY>("ROOT_QUERY", "products", 0, null, {
+        category: 2
+      })
+    ],
+    cacheBefore: {
+      ROOT_QUERY: {
+        'products({"category":1})': ["Product;1"],
+        'products({"category":2})': ["Product;2"]
+      },
+      "Product;1": { id: 1, sortNo: 0 },
+      "Product;2": { id: 2, sortNo: 0 }
+    },
+    cacheAfter: {
+      ROOT_QUERY: {
+        'products({"category":1})': ["Product;1"],
+        'products({"category":2})': [null, "Product;2"]
+      },
+      "Product;1": { id: 1, sortNo: 0 },
+      "Product;2": { id: 2, sortNo: 0 }
+    }
   }
 ];