+++ CHANGELOG.md
@@ -7,53 +7,67 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+- Patches for fields with arguments [#8](https://github.com/dividab/gql-cache-patch/pull/8).
+
 ## [v0.10.0] - 2019-02-19
 
 ### Added
+
 - Make invalidate recursive optional [#5](https://github.com/dividab/gql-cache-patch/pull/5).
 
 ## [v0.9.0] - 2018-09-28
 
 ### Changed
+
 - Upgrade gql-cache dependency to version 0.9.0
 
 ## [v0.8.0] - 2018-09-27
 
 ### Changed
+
 - Upgrade gql-cache dependency to version 0.8.0
 
 ## [v0.7.0] - 2018-09-14
 
 ### Changed
+
 - Upgrade gql-cache dependency to version 0.7.0
 
 ## [v0.6.0] - 2018-09-14
 
 ### Changed
+
 - Upgrade gql-cache dependency to version 0.6.0
 
 ## [v0.5.0] - 2018-09-14
 
 ### Added
+
 Add invalidate recursive and invalidate entity functionality
 
 ## [v0.4.0] - 2018-08-24
 
 ### Added
+
 - Upgraded dependency to gql-cache.
 - Fix check for non-existend field.
 
 ## [v0.3.0] - 2018-08-23
 
 ### Added
+
 - Determistic behaviour for missing data, see [#1](https://github.com/dividab/gql-cache-patch/issues/1).
 
 ## [v0.2.0] - 2018-08-23
 
 ### Added
+
 - Defaults to `Entity` for the type parameter in the constructor functions.
 
 ## v0.1.0 - 2018-08-17
+
 ### Added
 
 - Initial version.